+++ GLView/Images/GLImageView.h
@@ -2,7 +2,7 @@
 //  GLImageView.h
 //
 //  GLView Project
-//  Version 1.3.3
+//  Version 1.3.4
 //
 //  Created by Nick Lockwood on 10/07/2011.
 //  Copyright 2011 Charcoal Design